+++ bin/plugins/datasource-plugin-example/plugin_main.py
@@ -12,7 +12,7 @@ http://localhost:4000/dataframe/query
     {
       "format": "arrow",
       "name": "hello",
-      "location": "example://hello.world",
+      "location": "example://hello-world",
       "pluginOptions": {
         "foo": {
           "bar": [1, 2, 3],
@@ -21,25 +21,10 @@ http://localhost:4000/dataframe/query
           }
         }
       }
-      "schema": [
-        {
-          "name": "foo",
-          "dataType": "String",
-          "nullable": true
-        },
-        {
-          "name": "bar",
-          "dataType": "Int64",
-          "nullable": true
-        }
-      ]
     }
   ],
   "query": {
-    "sql": "SELECT * FROM hello"
-  },
-  "response": {
-    "format": "json"
+    "sql": "SELECT foo, bar, col_map FROM hello"
   }
 }
 ```
@@ -47,7 +32,7 @@ http://localhost:4000/dataframe/query
 Debug Message:
 ```
 >> Python received arguments:
-format, authority, path: arrow auth /path
+format, authority, path, other args
 schema: foo: string, bar: int64
 kwargs: {'foo': {'bar': [1.0, 2.0, 3.0], 'baz': {'value': 'text1'}}}
 ```
@@ -57,16 +42,23 @@ Response:
 [
   {
     "foo": "hello - arrow",
-    "bar": 12345
+    "bar": 12345,
+    "col_map": {
+      "key1": 123.4
+    }
   },
   {
     "foo": "world",
-    "bar": 67890
+    "bar": 67890,
+    "col_map": {
+      "key2": 567.8
+    }
   }
 ]
 ```
 """
 
+import logging
 import pyarrow as pa
 
 
@@ -91,7 +83,9 @@ def main(response_format: str, authority: str, path: str, schema: pa.Schema, **k
     )
 
     if response_format == "json":
-        return '[\n{"foo":"hello - json","bar":12345},\n{"foo":"world","bar":67890}\n]\n'
+        return (
+            '[\n{"foo":"hello - json","bar":12345},\n{"foo":"world","bar":67890}\n]\n'
+        )
 
     elif response_format == "ndJson":
         return '{"foo":"hello - ndJson","bar":12345}\n{"foo":"world","bar":67890}'
@@ -100,21 +94,35 @@ def main(response_format: str, authority: str, path: str, schema: pa.Schema, **k
         # Apache Arrow Python bindings - https://arrow.apache.org/docs/python/index.html
 
         # creates Schema when `schema` is `None`
-        schema = schema if schema is not None else (
-            pa.schema(
-                [
-                    pa.field("foo", pa.string()),
-                    pa.field("bar", pa.int64())
-                ]
+        schema = (
+            schema
+            if schema is not None
+            else (
+                pa.schema(
+                    [
+                        pa.field("foo", pa.string()),
+                        pa.field("bar", pa.int64()),
+                        pa.field("col_vector", pa.list_(pa.int32())),
+                        pa.field(
+                            "col_map",
+                            pa.map_(pa.string(), pa.float32()),
+                        ),
+                    ]
+                )
             )
         )
 
         # prepares columnar data
         foo = pa.array(["hello - arrow", "world"])
         bar = pa.array([12345, 67890])
+        col_vector = pa.array([[1, 2, 3, 4, 5], [6, 7, 8, 9, 10]])
+        col_map = pa.array(
+            [[{"key": "key1", "value": 123.4}], [{"key": "key2", "value": 567.8}]],
+            type=pa.map_(pa.string(), pa.float32())
+        )
 
         # creates RecordBatch and return
-        return pa.record_batch([foo, bar], schema=schema)
+        return pa.record_batch([foo, bar, col_vector, col_map], schema=schema)
 
     else:
         raise ValueError("Unsupported format: " + response_format)