+++ lib/src/data_source/csv_file.rs
@@ -22,7 +22,7 @@ pub fn to_record_batch(
     let delimiter = options.delimiter.unwrap_or(',') as u8;
 
     let df_schema = if let Some(schema) = schema {
-        SchemaRef::new(schema.to_datafusion_schema())
+        SchemaRef::new(schema.to_arrow_schema())
     } else {
         let format = csv::reader::Format::default()
             .with_header(has_header)