+++ lib/src/data_source/nd_json_rest.rs
@@ -19,7 +19,7 @@ pub async fn to_record_batch(
         .await?;
 
     let schema_ref = SchemaRef::new(if let Some(schema) = schema {
-        schema.to_datafusion_schema()
+        schema.to_arrow_schema()
     } else {
         infer_schema::from_raw_json(&response, options)?
     });