+++ bin/plugins/datasource-plugin-example/CHANGELOG.md
@@ -1,9 +1,13 @@
 # Changelog
 
+## 1.2.0 (2024-04-15)
+
+* Example of complex columns, List and Map
+
 ## 1.1.0 (2023-08-13)
 
 * Supports arrow for response format
-  + Receives or not arrow schema definitions to plugin main
+    + Receives or not arrow schema definitions to plugin main
 
 ## 1.0.0 (2023-04-09)